Fix: cast FFmpeg readFile output to Uint8Array for Blob constructor

+++ bob-studio-web/components/BurnSubtitles.tsx
@@ -49,7 +49,7 @@ export default function BurnSubtitles({ videoUrl, srtContent, fileName }: Props)
         'output.mp4',
       ])
 
-      const data = await ffmpeg.readFile('output.mp4')
+      const data = await ffmpeg.readFile('output.mp4') as Uint8Array
       const blob = new Blob([data], { type: 'video/mp4' })
       setOutputUrl(URL.createObjectURL(blob))
       setStatus('done')
